requirements.txt
djangorestframework>=3.15.0,<3.15.2
python-decouple<=3.8
djoser>=2.3.0,<2.3.1
djangorestframework-simplejwt>=5.3.0,<5.4.0
django-cors-headers<=4.6.0
drf_yasg<=1.21.8